requirements.txt
discord
requests
datetime
beautifulsoup4==4.9.3
python-dotenv==0.10.1